# backend/src/enterprise/enterpriseRoutes.ts
import type { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
import { createRequestContextMiddleware, getRequestContext } from "../context/requestContextMiddleware.js";
import { authorizeRequest } from "../gateway/authz.js";
import { businessApiAuth, getBusinessApiContext, requireBusinessScope } from "./businessApiAuth.js";
import { enterpriseService, EnterpriseError } from "./enterpriseService.js";

function fail(error:unknown,reply:FastifyReply){if(error instanceof EnterpriseError)return reply.status(error.statusCode).send({ok:false,error:error.message});if(error instanceof Error&&"statusCode" in error)return reply.status((error as Error&{statusCode:number}).statusCode).send({ok:false,error:error.message});throw error;}
function params(request:FastifyRequest){return request.params as {clientId:string;externalOrderId:string;deliveryId:string;credentialId:string;subscriptionId:string;billId:string};}
function adminGuard(request:FastifyRequest,reply:FastifyReply){const authz=authorizeRequest(getRequestContext(request));if(!authz.ok){reply.status(authz.statusCode).send({ok:false,error:authz.message});return false;}return true;}

export async function registerEnterpriseRoutes(app:FastifyInstance):Promise<void>{
  const adminContext=createRequestContextMiddleware({requireCityCode:true});

  app.get("/api/internal/enterprise/clients",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,clients:await enterpriseService.listClients(getRequestContext(request))};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,client:await enterpriseService.createClient(getRequestContext(request),request.body)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/status",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,client:await enterpriseService.updateClientStatus(getRequestContext(request),params(request).clientId,request.body)};}catch(error){return fail(error,reply);}});
  app.get("/api/internal/enterprise/clients/:clientId/credentials",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,credentials:await enterpriseService.listCredentials(getRequestContext(request),params(request).clientId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/credentials",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.createCredential(getRequestContext(request),params(request).clientId,request.body)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/credentials/:credentialId/revoke",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.revokeCredential(getRequestContext(request),params(request).clientId,params(request).credentialId)};}catch(error){return fail(error,reply);}});
  app.get("/api/internal/enterprise/clients/:clientId/agreement-prices",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,agreementPrices:await enterpriseService.listAgreements(getRequestContext(request),params(request).clientId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/agreement-prices",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,agreementPrice:await enterpriseService.upsertAgreement(getRequestContext(request),params(request).clientId,request.body)};}catch(error){return fail(error,reply);}});
  app.get("/api/internal/enterprise/clients/:clientId/webhook-subscriptions",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,subscriptions:await enterpriseService.listSubscriptionsForAdmin(getRequestContext(request),params(request).clientId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/webhook-subscriptions",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.createSubscriptionForAdmin(getRequestContext(request),params(request).clientId,request.body)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/webhook-subscriptions/:subscriptionId/status",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.updateSubscriptionStatus(getRequestContext(request),params(request).clientId,params(request).subscriptionId,request.body)};}catch(error){return fail(error,reply);}});
  app.get("/api/internal/enterprise/clients/:clientId/webhook-deliveries",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,deliveries:await enterpriseService.listDeliveries(getRequestContext(request),params(request).clientId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/webhook-deliveries/:deliveryId/retry",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.retryDelivery(getRequestContext(request),params(request).clientId,params(request).deliveryId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/webhooks/run-once",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.runWebhookDeliveries(getRequestContext(request))};}catch(error){return fail(error,reply);}});
  app.get("/api/internal/enterprise/clients/:clientId/bills",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,bills:await enterpriseService.listBills(getRequestContext(request),params(request).clientId)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/bills",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,bill:await enterpriseService.createBill(getRequestContext(request),params(request).clientId,request.body)};}catch(error){return fail(error,reply);}});
  app.post("/api/internal/enterprise/clients/:clientId/bills/:billId/issue",{preHandler:adminContext},async(request,reply)=>{if(!adminGuard(request,reply))return;try{return {ok:true,...await enterpriseService.issueBill(getRequestContext(request),params(request).clientId,params(request).billId)};}catch(error){return fail(error,reply);}});

  app.post("/openapi/v1/orders",{preHandler:businessApiAuth},async(request,reply)=>{try{const context=getBusinessApiContext(request);requireBusinessScope(context,"enterprise:orders:write");return {ok:true,...await enterpriseService.createOrder(context,request.body)};}catch(error){return fail(error,reply);}});
  app.get("/openapi/v1/orders",{preHandler:businessApiAuth},async(request,reply)=>{try{const context=getBusinessApiContext(request);requireBusinessScope(context,"enterprise:orders:read");return {ok:true,orders:await enterpriseService.listOrders(context)};}catch(error){return fail(error,reply);}});
  app.get("/openapi/v1/orders/:externalOrderId",{preHandler:businessApiAuth},async(request,reply)=>{try{const context=getBusinessApiContext(request);requireBusinessScope(context,"enterprise:orders:read");return {ok:true,order:await enterpriseService.getOrder(context,params(request).externalOrderId)};}catch(error){return fail(error,reply);}});
  app.get("/openapi/v1/webhook-subscriptions",{preHandler:businessApiAuth},async(request,reply)=>{try{const context=getBusinessApiContext(request);requireBusinessScope(context,"enterprise:webhooks:read");return {ok:true,subscriptions:await enterpriseService.listSubscriptionsForBusiness(context)};}catch(error){return fail(error,reply);}});
  app.post("/openapi/v1/webhook-subscriptions",{preHandler:businessApiAuth},async(request,reply)=>{try{const context=getBusinessApiContext(request);requireBusinessScope(context,"enterprise:webhooks:write");return {ok:true,...await enterpriseService.createSubscriptionForBusiness(context,request.body)};}catch(error){return fail(error,reply);}});
}