# docs/openapi/phase19-enterprise-v1.yaml
openapi: 3.1.0
info:
  title: XLB Enterprise Service API
  version: 1.0.0-phase19
  description: City and client scoped enterprise order and webhook subscription API.
servers:
  - url: https://api.example.invalid
    description: Replace with the deployed XLB API origin.
security:
  - BusinessApiKey: []
paths:
  /openapi/v1/orders:
    get:
      operationId: listEnterpriseOrders
      x-required-scope: enterprise:orders:read
      responses:
        "200": { description: Scoped enterprise order list }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createEnterpriseOrder
      x-required-scope: enterprise:orders:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEnterpriseOrder" }
      responses:
        "200": { description: Created or idempotently replayed enterprise order }
        "400": { description: Invalid request or unavailable SKU }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: External order id or idempotency conflict }
  /openapi/v1/orders/{externalOrderId}:
    get:
      operationId: getEnterpriseOrder
      x-required-scope: enterprise:orders:read
      parameters:
        - in: path
          name: externalOrderId
          required: true
          schema: { type: string, maxLength: 64 }
      responses:
        "200": { description: Enterprise order in the authenticated client and city }
        "404": { description: Order not found in client scope }
  /openapi/v1/webhook-subscriptions:
    get:
      operationId: listWebhookSubscriptions
      x-required-scope: enterprise:webhooks:read
      responses:
        "200": { description: Scoped subscription list without signing secrets }
    post:
      operationId: createWebhookSubscription
      x-required-scope: enterprise:webhooks:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWebhookSubscription" }
      responses:
        "200": { description: Subscription and one-time signing secret }
        "400": { description: Unsafe callback or unsupported event }
components:
  securitySchemes:
    BusinessApiKey:
      type: apiKey
      in: header
      name: X-XLB-API-Key
      description: One-time issued key in xlb.credential.secret format.
  responses:
    Unauthorized: { description: Missing, invalid, revoked, expired, or inactive-client API key }
    Forbidden: { description: API key lacks the required scope }
  schemas:
    CreateEnterpriseOrder:
      type: object
      additionalProperties: false
      required: [externalOrderId, idempotencyKey, skuId, quantity, addressProvince, addressCity, addressDistrict, detailAddress, contactName, contactPhone, scheduledAt, scheduledTimeSlot]
      properties:
        externalOrderId: { type: string, minLength: 2, maxLength: 64 }
        idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
        skuId: { type: string, maxLength: 128 }
        quantity: { type: integer, minimum: 1, maximum: 999 }
        addressProvince: { type: string, maxLength: 64 }
        addressCity: { type: string, maxLength: 64 }
        addressDistrict: { type: string, maxLength: 64 }
        detailAddress: { type: string, maxLength: 255 }
        contactName: { type: string, maxLength: 64 }
        contactPhone: { type: string, pattern: "^1[3-9][0-9]{9}$" }
        scheduledAt: { type: string, format: date-time }
        scheduledTimeSlot: { type: string, enum: [morning, afternoon, evening] }
    CreateWebhookSubscription:
      type: object
      additionalProperties: false
      required: [callbackUrl, eventTypes]
      properties:
        callbackUrl:
          type: string
          description: HTTPS for external delivery; mock scheme is local verification only.
        eventTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [order.created, order.paid, fulfillment.started, fulfillment.completed, fulfillment.evidence.created, fulfillment.customer_confirmation.confirmed, fulfillment.customer_confirmation.disputed, aftersale.complaint.submitted, aftersale.complaint.resolved]